Move getFuncDefByName to utils file for unit tests
