+++ src/providers/signatureProvider.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { Parser } from '../parser/parser';
+import { getFuncDefByName } from '../parser/parser.utils';
 
 export class SignatureProvider implements vscode.SignatureHelpProvider {
     public async provideSignatureHelp(
@@ -26,7 +26,7 @@ export class SignatureProvider implements vscode.SignatureHelpProvider {
             document.getWordRangeAtPosition(funcDefCandidatePosition),
         );
 
-        const funcDef = await Parser.getFuncDefByName(document, word);
+        const funcDef = getFuncDefByName(document.uri.path, word);
         if (funcDef) {
             return {
                 activeSignature: 0,
